… woltka/tests/data/output/bowtie2.free.tsv
… woltka/tests/test_align.py
… woltka/tests/test_classify.py
… woltka/tests/test_ordinal.py
… woltka/tests/test_workflow.py