Previously, the two mappers "plain" and "ordinal" were both Python classes which provide consistent interface. This is great but not performance-friendly, because class methods need to be called for every line in huge alignment files and class attributes need to be re-assigned. Replacing classes with native Python functions makes the code less pretty but meanwhile significantly improves performance. Benchmarks:

- Plain: 304 ms ± 1.82 ms => 264 ms ± 579 µs
- Ordinal: 928 ms ± 2.13 ms => 903 ms ± 4.62 ms